WIP: Allow initialization of the cluster from backups
Currently, replica's will already use the S3 backups if they are
available to create a $PGDATA directory.
However, if no master (endpoint) is available, the behaviour up to now
was to always do an initdb.

This may not be what the user wants, say - a user wants to suspend a
service - removing all the Kubernetes objects and only keeping the S3
backups.

In such a scenario, it would be great if the backup would be used to
initialize the whole cluster.